refactor(api): 검증 함수 분리
